requirements.txt
click>=8.2.1
matplotlib>=3.10.6
numba>=0.57.0
numpy>=2.2.6
gradio
Pillow
